Add support for logging to a file (while still printing log messages to the screen).
Run major commands by calling the "run" function.  This is a wrapper that captures
each program's output to a log file and simulates running of the command rather than
running it if $simulate is true.  (Still doesn't work properly if the command
contains redirection.  Also causes an extra level of shell expansion that may be
undesireable.)