Add support for logging to the system log if the shell variable syslog is set.
Add a warn() function to log a warning message.

Add a get_script_name() to determine the name of the script
currently running.  At the moment this just returns the base name
of $0, which should work provided the user uses POSIX style
functions and not KornShell style functions, i.e.:
    foo()
and not:
    function foo