Send an SMS message if a log is generated at notice priority. (e.g. if the checkdns.primary script changes the IP address  for our domain)
Use the common function library (use common versions of
functions like error and warn instead of the redundant ones
in checkdns.primary.)

Allow nameserver to be empty.  This will use the system's
default nameserver configuration.  In general this is a good
idea, but it won't work if the host doing the test has split
DNS (e.g. for endbracket.net, the internal DNS servers return
a 192.168.1.x address, but we are interested in the external
IP address.  In this case the user should define the
preferred nameserver using the following line in ~/.checkdns:
    nameserver=<IP address>