Add some usage information, just in case somebody else finds it and decides to use it.
Make the configuration and cache files hidden files in the user's
home directory.  I'm assuming the user that's running this is not
privileged, so the only directory they have read/write access to
may be their home directory, and I don't want to go to the effort
of creating a subdirectory in the user's home directory when a
hidden file is the standard UNIX way and will do just fine.

M    checkdns.backup